requirements.txt
celery[redis]
django
django-celery
django-registration
docker-py
redis
